backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
pydantic-settings==2.7.1
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
bcrypt==4.0.1
aiosmtplib==3.0.2
psycopg2-binary==2.9.10
fastmcp>=3.0.0
build>=0.10.0
toml>=0.10.0
asyncssh>=2.14.0
